requirements.txt
fastapi==0.111.1
pyjwt[crypto]==2.8.0